main entry point for the webapp
When ran, it opens the main page and runs all the flask apps as subprocesses. It also makes the urls look pretty. This runs on port 5000, so the other flask apps have to run on other ports. Additionally, mainPage.html, ajax.js, availableGamesScript,js need a few tweaks for it to work.